+++ common/constants.h
@@ -16,8 +16,8 @@
 // Software version
 #define FW_VERSION_MAJOR 1
 #define FW_VERSION_MINOR 1
-#define FW_VERSION_PATCH 2
-#define FW_VERSION_BUILD 1
+#define FW_VERSION_PATCH 3
+#define FW_VERSION_BUILD 0
 
 // Number of max sensors connected to the board
 #define CONNECTED_SENSORS 8