+++ sensors/sunrise/sunrise.c
@@ -115,6 +115,8 @@
 #define REG_ABC_PRESSURE_H          0xDE
 #define REG_ABC_PRESSURE_L          0xDF
 
+extern ms5607_t ms5607;
+
 /**
  * @brief returns error code according to the error register value
  * 
@@ -229,6 +231,14 @@ void sunrise_get_value(sensor_t* sunrise)
                 sunrise->temperature = NAN; // Set temperature to unknown
                 sunrise->state = SUNRISE_ERROR_WRONG_MODE; // Output Wrong mode error state
                 return;
+            }
+            if (sunrise->config.enable_pressure_comp && !ms5607.state) // Write pressure if enabled
+            {
+                uint8_t data[2];
+                sunrise->config.pressure = ms5607.pressure; // Copy pressure value
+                uint16_t val = ntoh16((uint16_t)(sunrise->config.pressure * 10));
+                memcpy(data, &val, 2);
+                sunrise_write(REG_AIR_PRESSURE_H, data, 2); // Write pressure from external probe
             }
             if (sunrise->config.single_meas_mode) // If in single measurement mode
             {